change to roothash chain

+++ packages/contracts-theweb3chain/src/libraries/BN254.sol
@@ -28,15 +28,12 @@ pragma solidity ^0.8.24;
  */
 library BN254 {
     // modulus for the underlying field F_p of the elliptic curve
-    uint256 internal constant FP_MODULUS =
-        21888242871839275222246405745257275088696311157297823662689037894645226208583;
+    uint256 internal constant FP_MODULUS = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
     // modulus for the underlying field F_r of the elliptic curve
-    uint256 internal constant FR_MODULUS =
-        21888242871839275222246405745257275088548364400416034343698204186575808495617;
+    uint256 internal constant FR_MODULUS = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
 
-    // primitive root of unity 
+    // primitive root of unity
     uint256 internal constant OMEGA = 10359452186428527605436343203440067497552205259388878191021578220384701716497;
-    
 
     struct G1Point {
         uint256 X;
@@ -51,39 +48,29 @@ library BN254 {
 
     // generator of group G2
     /// @dev Generator point in F_q2 is of the form: (x0 + ix1, y0 + iy1).
-    uint256 internal constant G2x1 =
-        11559732032986387107991004021392285783925812861821192530917403151452391805634;
-    uint256 internal constant G2x0 =
-        10857046999023057135944570762232829481370756359578518086990519993285655852781;
-    uint256 internal constant G2y1 =
-        4082367875863433681332203403145435568316851327593401208105741076214120093531;
-    uint256 internal constant G2y0 =
-        8495653923123431417604973247489272438418190587263600148770280649306958101930;
+    uint256 internal constant G2x1 = 11559732032986387107991004021392285783925812861821192530917403151452391805634;
+    uint256 internal constant G2x0 = 10857046999023057135944570762232829481370756359578518086990519993285655852781;
+    uint256 internal constant G2y1 = 4082367875863433681332203403145435568316851327593401208105741076214120093531;
+    uint256 internal constant G2y0 = 8495653923123431417604973247489272438418190587263600148770280649306958101930;
     /// @notice returns the G2 generator
     /// @dev mind the ordering of the 1s and 0s!
     ///      this is because of the (unknown to us) convention used in the bn254 pairing precompile contract
     ///      "Elements a * i + b of F_p^2 are encoded as two elements of F_p, (a, b)."
     ///      https://github.com/ethereum/EIPs/blob/master/EIPS/eip-197.md#encoding
+
     function generatorG2() internal pure returns (G2Point memory) {
-        return G2Point(
-            [G2x1, G2x0], [G2y1, G2y0]
-        );
+        return G2Point([G2x1, G2x0], [G2y1, G2y0]);
     }
 
     // negation of the generator of group G2
     /// @dev Generator point in F_q2 is of the form: (x0 + ix1, y0 + iy1).
-    uint256 internal constant nG2x1 =
-        11559732032986387107991004021392285783925812861821192530917403151452391805634;
-    uint256 internal constant nG2x0 =
-        10857046999023057135944570762232829481370756359578518086990519993285655852781;
-    uint256 internal constant nG2y1 =
-        17805874995975841540914202342111839520379459829704422454583296818431106115052;
-    uint256 internal constant nG2y0 =
-        13392588948715843804641432497768002650278120570034223513918757245338268106653;
+    uint256 internal constant nG2x1 = 11559732032986387107991004021392285783925812861821192530917403151452391805634;
+    uint256 internal constant nG2x0 = 10857046999023057135944570762232829481370756359578518086990519993285655852781;
+    uint256 internal constant nG2y1 = 17805874995975841540914202342111839520379459829704422454583296818431106115052;
+    uint256 internal constant nG2y0 = 13392588948715843804641432497768002650278120570034223513918757245338268106653;
+
     function negGeneratorG2() internal pure returns (G2Point memory) {
-        return G2Point(
-            [nG2x1, nG2x0], [nG2y1, nG2y0]
-        );
+        return G2Point([nG2x1, nG2x0], [nG2y1, nG2y0]);
     }
 
     // first power of srs in G2
@@ -92,15 +79,12 @@ library BN254 {
     uint256 internal constant G2SRSx0 = 21039730876973405969844107393779063362038454413254731404052240341412356318284;
     uint256 internal constant G2SRSy1 = 18697407556011630376420900106252341752488547575648825575049647403852275261247;
     uint256 internal constant G2SRSy0 = 7586489485579523767759120334904353546627445333297951253230866312564920951171;
+
     function G2SRSFirstPower() internal pure returns (G2Point memory) {
-        return G2Point(
-            [G2SRSx0, G2SRSx1], [G2SRSy0, G2SRSy1]
-        );
+        return G2Point([G2SRSx0, G2SRSx1], [G2SRSy0, G2SRSy1]);
     }
 
-    bytes32 internal constant powersOfTauMerkleRoot =
-        0x22c998e49752bbb1918ba87d6d59dd0e83620a311ba91dd4b2cc84990b31b56f;
-
+    bytes32 internal constant powersOfTauMerkleRoot = 0x22c998e49752bbb1918ba87d6d59dd0e83620a311ba91dd4b2cc84990b31b56f;
 
     /**
      * @param p Some point in G1.
@@ -118,10 +102,7 @@ library BN254 {
     /**
      * @return r the sum of two points of G1
      */
-    function plus(
-        G1Point memory p1,
-        G1Point memory p2
-    ) internal view returns (G1Point memory r) {
+    function plus(G1Point memory p1, G1Point memory p2) internal view returns (G1Point memory r) {
         uint256[4] memory input;
         input[0] = p1.X;
         input[1] = p1.Y;
@@ -134,9 +115,7 @@ library BN254 {
             success := staticcall(sub(gas(), 2000), 6, input, 0x80, r, 0x40)
             // Use "invalid" to make gas estimation work
             switch success
-            case 0 {
-                invalid()
-            }
+            case 0 { invalid() }
         }
 
         require(success, "ec-add-failed");
@@ -147,10 +126,7 @@ library BN254 {
      *         p == p.scalar_mul(1) and p.plus(p) == p.scalar_mul(2) for all
      *         points p.
      */
-    function scalar_mul(
-        G1Point memory p,
-        uint256 s
-    ) internal view returns (G1Point memory r) {
+    function scalar_mul(G1Point memory p, uint256 s) internal view returns (G1Point memory r) {
         uint256[3] memory input;
         input[0] = p.X;
         input[1] = p.Y;
@@ -161,9 +137,7 @@ library BN254 {
             success := staticcall(sub(gas(), 2000), 7, input, 0x60, r, 0x40)
             // Use "invalid" to make gas estimation work
             switch success
-            case 0 {
-                invalid()
-            }
+            case 0 { invalid() }
         }
         require(success, "ec-mul-failed");
     }
@@ -174,12 +148,11 @@ library BN254 {
      *         For example,
      *         pairing([P1(), P1().negate()], [P2(), P2()]) should return true.
      */
-    function pairing(
-        G1Point memory a1,
-        G2Point memory a2,
-        G1Point memory b1,
-        G2Point memory b2
-    ) internal view returns (bool) {
+    function pairing(G1Point memory a1, G2Point memory a2, G1Point memory b1, G2Point memory b2)
+        internal
+        view
+        returns (bool)
+    {
         G1Point[2] memory p1 = [a1, b1];
         G2Point[2] memory p2 = [a2, b2];
 
@@ -200,19 +173,10 @@ library BN254 {
 
         // solium-disable-next-line security/no-inline-assembly
         assembly {
-            success := staticcall(
-                sub(gas(), 2000),
-                8,
-                input,
-                mul(12, 0x20),
-                out,
-                0x20
-            )
+            success := staticcall(sub(gas(), 2000), 8, input, mul(12, 0x20), out, 0x20)
             // Use "invalid" to make gas estimation work
             switch success
-            case 0 {
-                invalid()
-            }
+            case 0 { invalid() }
         }
 
         require(success, "pairing-opcode-failed");
@@ -224,13 +188,11 @@ library BN254 {
      * @notice This function is functionally the same as pairing(), however it specifies a gas limit
      *         the user can set, as a precompile may use the entire gas budget if it reverts.
      */
-    function safePairing(
-        G1Point memory a1,
-        G2Point memory a2,
-        G1Point memory b1,
-        G2Point memory b2,
-        uint256 pairingGas
-    ) internal view returns (bool, bool) {
+    function safePairing(G1Point memory a1, G2Point memory a2, G1Point memory b1, G2Point memory b2, uint256 pairingGas)
+        internal
+        view
+        returns (bool, bool)
+    {
         G1Point[2] memory p1 = [a1, b1];
         G2Point[2] memory p2 = [a2, b2];
 
@@ -251,14 +213,7 @@ library BN254 {
 
         // solium-disable-next-line security/no-inline-assembly
         assembly {
-            success := staticcall(
-                pairingGas,
-                8,
-                input,
-                mul(12, 0x20),
-                out,
-                0x20
-            )
+            success := staticcall(pairingGas, 8, input, mul(12, 0x20), out, 0x20)
         }
 
         //Out is the output of the pairing precompile, either 0 or 1 based on whether the two pairings are equal.
@@ -269,12 +224,9 @@ library BN254 {
 
     /// @return the keccak256 hash of the G1 Point
     /// @dev used for BLS signatures
-    function hashG1Point(
-        BN254.G1Point memory pk
-    ) internal pure returns (bytes32) {
+    function hashG1Point(BN254.G1Point memory pk) internal pure returns (bytes32) {
         return keccak256(abi.encodePacked(pk.X, pk.Y));
     }
-
 
     /**
      * @notice adapted from https://github.com/HarryR/solcrypto/blob/master/contracts/altbn128.sol
@@ -286,11 +238,11 @@ library BN254 {
         // XXX: Gen Order (n) or Field Order (p) ?
         uint256 x = uint256(_x) % FP_MODULUS;
 
-        while( true ) {
+        while (true) {
             (beta, y) = findYFromX(x);
 
             // y^2 == beta
-            if( beta == mulmod(y, y, FP_MODULUS) ) {
+            if (beta == mulmod(y, y, FP_MODULUS)) {
                 return (x, y);
             }
 
@@ -300,15 +252,13 @@ library BN254 {
     }
 
     /**
-    * Given X, find Y
-    *
-    *   where y = sqrt(x^3 + b)
-    *
-    * Returns: (x^3 + b), y
-    */
-    function findYFromX(uint256 x)
-        internal view returns(uint256, uint256)
-    {
+     * Given X, find Y
+     *
+     *   where y = sqrt(x^3 + b)
+     *
+     * Returns: (x^3 + b), y
+     */
+    function findYFromX(uint256 x) internal view returns (uint256, uint256) {
         // beta = (x^3 + b) % p
         uint256 beta = addmod(mulmod(mulmod(x, x, FP_MODULUS), x, FP_MODULUS), 3, FP_MODULUS);
 
@@ -322,17 +272,18 @@ library BN254 {
     function expMod(uint256 _base, uint256 _exponent, uint256 _modulus) internal view returns (uint256 retval) {
         bool success;
         uint256[1] memory output;
-        uint[6] memory input;
-        input[0] = 0x20;        // baseLen = new(big.Int).SetBytes(getData(input, 0, 32))
-        input[1] = 0x20;        // expLen  = new(big.Int).SetBytes(getData(input, 32, 32))
-        input[2] = 0x20;        // modLen  = new(big.Int).SetBytes(getData(input, 64, 32))
+        uint256[6] memory input;
+        input[0] = 0x20; // baseLen = new(big.Int).SetBytes(getData(input, 0, 32))
+        input[1] = 0x20; // expLen  = new(big.Int).SetBytes(getData(input, 32, 32))
+        input[2] = 0x20; // modLen  = new(big.Int).SetBytes(getData(input, 64, 32))
         input[3] = _base;
         input[4] = _exponent;
         input[5] = _modulus;
         assembly {
             success := staticcall(sub(gas(), 2000), 5, input, 0xc0, output, 0x20)
             // Use "invalid" to make gas estimation work
-            switch success case 0 { invalid() }
+            switch success
+            case 0 { invalid() }
         }
         require(success);
         return output[0];
